Started quote functionality, minor changes
.quote - replies a random quote from the list
.quoteid id - replies the quote with the given id
.addquote string - adds string to the quote list

+++ main.py
@@ -1,12 +1,12 @@
-
 import sys
-import discord
-import config
-from debug import dbprint
-from TwitterAPI import TwitterAPI
-from discord.ext import commands
-from colorama import Fore
 import random
+import config
+import discord
+import pymysql.cursors
+from colorama import Fore
+from debug import dbprint
+from discord.ext import commands
+from TwitterAPI import TwitterAPI
 
 api = TwitterAPI(config.appKey, config.appSecret, config.oauthToken, config.oauthTokenSecret)
 description = '''A learning experience bot, written in python using discory.py .'''
@@ -21,11 +21,9 @@ async def on_ready():
     print(Fore.WHITE + '________________________')
     dbprint('info', sys.platform)
     if sys.platform != "win32":
-        await bot.change_presence(game=discord.Game(name='!info'))
-        dbprint('info', '!info')
+        await bot.change_presence(game=discord.Game(name='.info'))
     else:
-        await bot.change_presence(game=discord.Game(name='!info - indev'))
-        dbprint('info', '!info - indev')
+        await bot.change_presence(game=discord.Game(name='.info - indev'))
     print(Fore.GREEN + "connected successfully")
     print(Fore.WHITE + '________________________')
 
@@ -49,6 +47,7 @@ async def twitch(ctx):
     await bot.delete_message(ctx.message)
     await bot.send_typing(ctx.message.channel)
     await bot.send_message(ctx.message.channel, "<https://twitch.tv/thebloodyscreen>")
+# simple shit above
 
 
 # everything voice --begin
@@ -87,10 +86,10 @@ async def leave():
 # simple twitter post command
 @bot.command(pass_context=True)
 async def tweet(ctx):
+    await bot.delete_message(ctx.message)
     if str(ctx.message.author) == 'TheBloodyScreen#4278':
         r = api.request('statuses/update', {'status': ctx.message.content.replace('.tweet ', '')})
-        await bot.say('SUCCESS' if r.status_code == 200 else 'FAILURE')
-
+        await bot.send_message(ctx.message.author, 'SUCCESS' if r.status_code == 200 else 'FAILURE')
     else:
         await bot.say("You don't have permission to use this command!")
 
@@ -111,21 +110,99 @@ async def choose(*choices: str):
 @bot.command(pass_context=True)
 async def repeat(ctx, times: int, *, content: str):
     await bot.delete_message(ctx.message)
-    if 'amins' in getRoles(ctx.message.author):
+    if 'admins' in getRoles(ctx.message.author):
         for i in range(times):
             await bot.say(content)
     else:
         await bot.say("You don't have permission to use this command!")
 
 
+# It's like "the purge," Morty. That movie "the purge"?
+@bot.command(pass_context=True)
+async def purge(ctx, amount):
+    if 'admins' in getRoles(ctx.message.author):
+        try:
+            await bot.delete_message(ctx.message)
+            await bot.purge_from(ctx.message.channel, limit=int(amount))
+        except:
+            await bot.delete_message(ctx.message)
+            await bot.say(ctx.message.channel, "Please provide an int!")
+    else:
+        await bot.delete_message(ctx.message)
+        await bot.say(ctx.message.channel, "You don't have permission to use this command!")
+
+
+# show me what you got
+@bot.command(pass_context=True)
+async def roles(ctx):
+    for server in bot.servers:
+        for member in server.members:
+            if member.mentioned_in(ctx.message):
+                user = member
+    await bot.send_message(ctx.message.channel, embed=discord.Embed(title=str(str(user)[0:-5]) + ' has the following Groups:', description=("\n".join(getRoles(user))), color=0x0000FF))
+
+
+# tell me something fun
+@bot.command(pass_context=True)
+async def quote(ctx):
+    await bot.delete_message(ctx.message)
+    connection = pymysql.connect(host=config.host, user=config.user, password=config.password, db=config.db, charset=config.charset, cursorclass=pymysql.cursors.DictCursor)
+
+    try:
+        with connection.cursor() as cursor:
+            sql = "SELECT TABLE_NAME AS 'Table Name', TABLE_ROWS AS 'Rows' FROM information_schema.TABLES WHERE TABLES.TABLE_SCHEMA = 'TheBloodyDiscord' AND TABLES.TABLE_TYPE = 'BASE TABLE'"
+            cursor.execute(sql)
+            if sys.platform != "win32":
+                maxid = (str(cursor.fetchall()[0]).replace("{'Table Name': 'quotes', 'Rows': ", '')).replace('}', '')
+            else:
+                maxid = (str(cursor.fetchall()[0]).replace("{'Table Name': 'quotes', 'Rows': ", '')).replace('}', '')
+            sql = "SELECT `quote` FROM `quotes` WHERE `id`=%s"
+            cursor.execute(sql, (random.randint(1, int(maxid))))
+            result = (str(cursor.fetchone()).replace("{'quote': ", '')).replace('}', '')
+            dbprint('info', result)
+            await bot.say(result)
+    finally:
+        connection.close()
+
+
+@bot.command(pass_context=True)
+async def quoteid(ctx):
+    await bot.delete_message(ctx.message)
+    connection = pymysql.connect(host=config.host, user=config.user, password=config.password, db=config.db, charset=config.charset, cursorclass=pymysql.cursors.DictCursor)
+
+    try:
+        with connection.cursor() as cursor:
+            sql = "SELECT `quote` FROM `quotes` WHERE `id`=%s"
+            cursor.execute(sql, int(str(ctx.message.content).replace('.quoteid ', '')))
+            result = (str(cursor.fetchone()).replace("{'quote': '", '')).replace("'}", "")
+            await bot.say(result)
+    finally:
+        connection.close()
+
+
+@bot.command(pass_context=True)
+async def addquote(ctx):
+    await bot.delete_message(ctx.message)
+    connection = pymysql.connect(host=config.host, user=config.user, password=config.password, db=config.db, charset=config.charset, cursorclass=pymysql.cursors.DictCursor)
+
+    try:
+        with connection.cursor() as cursor:
+            sql = "INSERT INTO `quotes` (`quote`) VALUES (%s)"
+            message = str(ctx.message.content).replace('.addquote ', '')
+            cursor.execute(sql, message,)
+        connection.commit()
+    finally:
+        connection.close()
+
+
 # markdown helper
 @bot.event
 async def on_message(message):
     author = message.author
-    if message.content.startswith('CODE') and author.id != '322081774622605312':
+    if message.content.startswith('CODE') and message.author != 'TheBloodyDiscord#6405' and message.author.id != '322081774622605312':
+        author = str(message.author)
         await bot.delete_message(message)
         newMessage = str(message.content).replace("CODE", "```")
-        author = str(message.author)
         await bot.send_message(message.channel, "CODE by " + str(author[0:-5]) + newMessage + "\n```")
     await bot.process_commands(message)
